+++ src/index.ts
@@ -30,6 +30,7 @@ export * from "./widgets/growable-items-container/growable-items-container";
 export * from "./utils/use-query-params/use-query-params";
 export * from "./utils/use-countdown/use-countdown";
 export * from "./utils/use-window-breakpoints/use-window-breakpoints";
+export * from "./utils/debounce/debounce";
 
 export * from "../src-old/form/file-picker";
 export * from "../src-old/form/file-field";
@@ -37,5 +38,3 @@ export * from "../src-old/form/date-picker";
 export * from "../src-old/form/date-field";
 export * from "../src-old/form/time-picker";
 export * from "../src-old/form/time-field";
-
-export * from "../src-old/misc/debounce";