+++ src/utils/debounce/debounce.ts
@@ -1,4 +1,6 @@
-export const debounce = (label: string, callback: Function, delay: number): void => {
+import { Debounce } from "../../../types";
+
+export const debounce: Debounce = (label: string, callback: Function, delay: number): void => {
     const w: any = window;
     const timerKey = `debounced-function-${label}`;
     if (w[timerKey]) clearTimeout(w[timerKey]);