+++ CHANGELOG.md
@@ -76,7 +76,7 @@
 
 -   `1.7.3`
 
-    -   Added [debounce](src-old/misc/debounce/usage.md) utility
+    -   Added [debounce](src/utils/debounce/usage.md) utility
     -   Added [useCountdown](src/utils/use-countdown/use-countdown-usage.md) hook
     -   Added `initialStorageKeys` prop to [ValueStoreProvider](src/widgets/local-storage-provider/local-storage-provider-usage.md)
         to allow pre-existing local storage values to be added to the value store